exit 1

+++ Tools/create_fontbnd/create_fontbnd.sh
@@ -1,5 +1,16 @@
 scripts/convert_dds.sh
+if [ $? -eq 1 ]; then
+    exit 1
+fi
+
 scripts/create_tpf.sh
+if [ $? -eq 1 ]; then
+    exit 1
+fi
+
 scripts/create_bnd4.sh
+if [ $? -eq 1 ]; then
+    exit 1
+fi
 
 rm -r "tmp/"
